+++ tests/Handler/HeaderProcessorTest.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace GuzzleHttp\Tests\Handler;
+
+use GuzzleHttp\Handler\HeaderProcessor;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \GuzzleHttp\Handler\HeaderProcessor
+ */
+class HeaderProcessorTest extends TestCase
+{
+    public function testParsesLastHeaderBlock(): void
+    {
+        [$version, $status, $reason, $headers] = HeaderProcessor::parseHeaders([
+            'HTTP/1.1 100 Continue',
+            'Ignored: header',
+            'HTTP/1.1 200 OK',
+            'X-Foo: bar',
+        ]);
+
+        self::assertSame('1.1', $version);
+        self::assertSame(200, $status);
+        self::assertSame('OK', $reason);
+        self::assertSame(['X-Foo' => ['bar']], $headers);
+    }
+
+    public function testRejectsMalformedStatusCode(): void
+    {
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('HTTP status code is invalid');
+
+        HeaderProcessor::parseHeaders([
+            'HTTP/1.1 200abc Weird',
+        ]);
+    }
+}